PawanPro/Flask Bot/requirements.txt
Flask==3.0.0
PyPDF2==3.0.1
python-docx==1.0.1
requests==2.31.0
Werkzeug==3.0.1
python-dotenv==1.0.0
gunicorn==21.2.0
Flask-Cors==4.0.0
Pillow==10.2.0
pytest==7.4.4
black==24.1.1
flake8==7.0.0
pytest-cov==4.1.0

Flask==2.3.3
PyPDF2==3.0.1
python-docx==0.8.11
Werkzeug==2.3.7
requests==2.31.0
python-dotenv==1.0.0